fix(group): Make groups cluster scoped
We do not actually need to make the groups part of namespaces, they can
be cluster scoped, this way any ApiCheck in any namespace can read it.
It's a much better solution than specifying the namespace for each
group.

Tested locally, tests updated to not use a namespace for the group
resource.